+++ src/plugins/piral-notifications/src/actions.ts
@@ -1,4 +1,4 @@
-import { prependItem, excludeItem, GlobalStateContext } from 'piral-core';
+import { prependItem, excludeOn, GlobalStateContext } from 'piral-core';
 import { OpenNotification } from './types';
 
 export function openNotification(ctx: GlobalStateContext, dialog: OpenNotification) {
@@ -11,6 +11,6 @@ export function openNotification(ctx: GlobalStateContext, dialog: OpenNotificati
 export function closeNotification(ctx: GlobalStateContext, dialog: OpenNotification) {
   ctx.dispatch((state) => ({
     ...state,
-    notifications: excludeItem(state.notifications, dialog),
+    notifications: excludeOn(state.notifications, (notification) => notification.id === dialog.id),
   }));
 }