+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.13.1 (tbd)
 
+* Fixed closing notifications after modifications in `piral-notifications`
+* Fixed closing modal dialogs after modifications in `piral-modals`
 * Implemented handling for navigation in `piral-blazor`
 * Updated development dependencies
 * Updated `tslib` to v2